style: PrevButton을 공통 컴포넌트로 분리
