+++ front/src/components/Letter/LetterListHeader/LetterListHeader.tsx
@@ -1,21 +1,5 @@
-import { useNavigate } from 'react-router-dom';
 import PageTitle from '../../Common/PageTitle/PageTitle';
-import { ReactComponent as PrevButtonIcon } from '../../../assets/img/prev_button.svg';
-
-const PrevButton = () => {
-  const naviagte = useNavigate();
-
-  const onClickHandler = (): void => {
-    console.log('뒤로가기 버튼');
-    naviagte(-1);
-  };
-
-  return (
-    <button onClick={onClickHandler}>
-      <PrevButtonIcon />
-    </button>
-  );
-};
+import PrevButton from '../../Common/PrevButton/PrevButton';
 
 const LetterListHeader = () => {
   return (
